Readme
This repository contains data and visualizations related to the world population. The data is sourced from Kaggle on the 10th of October 2022 and includes information on population by country, age and gender.

## Data
The data is provided in CSV format and includes the following columns:
the rank by population
three-digit country/territories code (CCA3)
the name of the countries/territories (country)
the Capital (Capital)
the name of the continent (Continent)
Population of the country in the year 2022 (2022 Population)
Population of the country in the year 2020 (2020 Population)
Population of the country in the year 2015 (2015 Population)
Population of the country in the year 2010 (2010 Population)
Population of the country in the year 2000 (2000 Population)
Population of the country in the year 1990 (1990 Population)
Population of the country in the year 1980 (1980 Population)
Population of the country in the year 1970 (1970 Population)
Area size of the Country/Territories in square kilometer (Area(km2))
Population Growth rate by Countries/Territories (Growth Rate)
Population percentage by each country/Territories (World Population Percentage)
		
## Visualizations
The repository includes several visualizations of the data, created using ggplot2 and plotly from Rstudio. These visualizations include Barplots of population for selected countries, age group etc. Pie charts of population for selected countries grouped by gender and Interactive map of world population.